♻️ Refactor: 파일 폴더 구조 리펙토링
