+++ src/features/dashboards/apis/getMyDashboards.ts
@@ -1,6 +1,6 @@
 import type { DashboardListResponse } from '@/features/dashboards/apis/dashboards.types';
-import { getDashboardColorName } from '@/features/dashboards/constants/dashboardColorMap.constants';
 import type { DashboardItem } from '@/features/dashboards/types/myDashboard.types';
+import { getDashboardColorName } from '@/features/dashboards/utils/dashboardColor';
 import { COLORS } from '@/shared/constants/color.constants';
 import { get } from '@/shared/apis/fetchInstance';
