+++ src/pages/my-dashboard/index.tsx
@@ -2,8 +2,8 @@ import { Link } from 'react-router-dom';
 import { IcArrowRight } from '@/shared/assets/icons';
 import CreateDashboardModal from '@/features/dashboards/components/create-dashboard-modal';
 import { ColorLabel } from '@/features/dashboards/components/color/color-label';
-import { getDashboardColorHex } from '@/features/dashboards/constants/dashboardColorMap.constants';
 import { useDashboardList } from '@/features/dashboards/hooks/useDashboardList';
+import { getDashboardColorHex } from '@/features/dashboards/utils/dashboardColor';
 import { Button } from '@/shared/components/button';
 import { PageIndicator } from '@/shared/components/page-indicator';
 import AddDashboardButton from '@/pages/my-dashboard/components/add-dashboard-button';
